+++ src/jp/ac/kobe_u/cs/sugar/csp/IntegerDomain.java
@@ -15,17 +15,17 @@ import jp.ac.kobe_u.cs.sugar.expression.Expression;
  * @author Naoyuki Tamura (tamura@kobe-u.ac.jp)
  */
 public class IntegerDomain {
-	public static int MAX_SET_SIZE = 128;
-	private int lb;
-	private int ub;
-	private SortedSet<Integer> domain;
+	public static long MAX_SET_SIZE = 128;
+	private long lb;
+	private long ub;
+	private SortedSet<Long> domain;
 
-	private static IntegerDomain create(SortedSet<Integer> domain) throws SugarException {
-		int lb = domain.first();
-		int ub = domain.last();
+	private static IntegerDomain create(SortedSet<Long> domain) throws SugarException {
+		long lb = domain.first();
+		long ub = domain.last();
 		if (domain.size() <= MAX_SET_SIZE) {
 			boolean sparse = false;
-			for (int value = lb; value <= ub; value++) {
+			for (long value = lb; value <= ub; value++) {
 				if (! domain.contains(value)) {
 					sparse = true;
 					break;
@@ -43,7 +43,7 @@ public class IntegerDomain {
 		return new IntegerDomain(domain);
 	}
 
-	public IntegerDomain(int lb, int ub) throws SugarException {
+	public IntegerDomain(long lb, long ub) throws SugarException {
 		if (lb > ub) {
 			throw new SugarException("Illegal domain instantiation " + lb + " " + ub);
 		}
@@ -52,7 +52,7 @@ public class IntegerDomain {
 		domain = null;
 	}
 
-	public IntegerDomain(SortedSet<Integer> domain) {
+	public IntegerDomain(SortedSet<Long> domain) {
 		lb = domain.first();
 		ub = domain.last();
 		this.domain = domain;
@@ -63,15 +63,15 @@ public class IntegerDomain {
 		ub = d.ub;
 		domain = null;
 		if (d.domain != null) {
-			domain = new TreeSet<Integer>(d.domain);
+			domain = new TreeSet<Long>(d.domain);
 		}
 	}
 
 	private IntegerDomain() {
-		domain = new TreeSet<Integer>();
+		domain = new TreeSet<Long>();
 	}
 
-	public int size() {
+	public long size() {
 		if (domain == null) {
 			return lb <= ub ? ub - lb + 1 : 0;
 		} else {
@@ -87,7 +87,7 @@ public class IntegerDomain {
 		return size() == 0;
 	}
 
-	public int getLowerBound() {
+	public long getLowerBound() {
 		if (domain == null) {
 			return lb;
 		} else {
@@ -95,7 +95,7 @@ public class IntegerDomain {
 		}
 	}
 
-	public int getUpperBound() {
+	public long getUpperBound() {
 		if (domain == null) {
 			return ub;
 		} else {
@@ -103,7 +103,7 @@ public class IntegerDomain {
 		}
 	}
 
-	public boolean contains(int value) {
+	public boolean contains(long value) {
 		if (domain == null) {
 			return lb <= value && value <= ub;
 		} else {
@@ -111,7 +111,7 @@ public class IntegerDomain {
 		}
 	}
 
-	public IntegerDomain bound(int lb, int ub) throws SugarException {
+	public IntegerDomain bound(long lb, long ub) throws SugarException {
 		if (lb <= this.lb && this.ub <= ub)
 			return this;
 		if (domain == null) {
@@ -123,11 +123,11 @@ public class IntegerDomain {
 		}
 	}
 
-	private class Iter implements Iterator<Integer> {
-		int value;
-		int ub;
+	private class Iter implements Iterator<Long> {
+		long value;
+		long ub;
 
-		public Iter(int lb, int ub) {
+		public Iter(long lb, long ub) {
 			value = lb;
 			this.ub = ub;
 		}
@@ -136,7 +136,7 @@ public class IntegerDomain {
 			return value <= ub;
 		}
 
-		public Integer next() {
+		public Long next() {
 			return value++;
 		}
 
@@ -144,7 +144,7 @@ public class IntegerDomain {
 		}
 	}
 
-	public Iterator<Integer> values(int lb, int ub) {
+	public Iterator<Long> values(long lb, long ub) {
 		if (lb > ub) {
 			return new Iter(lb, ub);
 		} else if (domain == null) {
@@ -156,17 +156,17 @@ public class IntegerDomain {
 		}
 	}
 
-	public Iterator<Integer> values() {
+	public Iterator<Long> values() {
 		return values(lb, ub);
 	}
 
 	public IntegerDomain cup(IntegerDomain d1) throws SugarException {
 		if (domain == null || d1.domain == null) {
-			int lb = Math.min(this.lb, d1.lb);
-			int ub = Math.max(this.ub, d1.ub);
+			long lb = Math.min(this.lb, d1.lb);
+			long ub = Math.max(this.ub, d1.ub);
 			return new IntegerDomain(lb, ub);
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>(domain);
+			SortedSet<Long> d = new TreeSet<Long>(domain);
 			d.addAll(d1.domain);
 			// return new IntegerDomain(d);
 			return create(d);
@@ -179,8 +179,8 @@ public class IntegerDomain {
 		} else if (domain == null) {
 			return d1.bound(lb, ub);
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>();
-			for (int value : domain) {
+			SortedSet<Long> d = new TreeSet<Long>();
+			for (long value : domain) {
 				if (d1.contains(value)) {
 					d.add(value);
 				}
@@ -193,8 +193,8 @@ public class IntegerDomain {
 		if (domain == null) {
 			return new IntegerDomain(-ub, -lb);
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>();
-			for (int value : domain) {
+			SortedSet<Long> d = new TreeSet<Long>();
+			for (long value : domain) {
 				d.add(-value);
 			}
 			// return new IntegerDomain(d);
@@ -204,16 +204,16 @@ public class IntegerDomain {
 
 	public IntegerDomain abs() throws SugarException {
 		if (domain == null) {
-			int lb0 = Math.min(Math.abs(lb), Math.abs(ub));
-			int ub0 = Math.max(Math.abs(lb), Math.abs(ub));
+			long lb0 = Math.min(Math.abs(lb), Math.abs(ub));
+			long ub0 = Math.max(Math.abs(lb), Math.abs(ub));
 			if (lb <= 0 && 0 <= ub) {
 				return new IntegerDomain(0, ub0);
 			} else {
 				return new IntegerDomain(lb0, ub0);
 			}
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>();
-			for (int value : domain) {
+			SortedSet<Long> d = new TreeSet<Long>();
+			for (long value : domain) {
 				d.add(Math.abs(value));
 			}
 			// return new IntegerDomain(d);
@@ -221,12 +221,12 @@ public class IntegerDomain {
 		}
 	}
 
-	public IntegerDomain add(int a) throws SugarException {
+	public IntegerDomain add(long a) throws SugarException {
 		if (domain == null) {
 			return new IntegerDomain(lb+a, ub+a);
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>();
-			for (int value : domain) {
+			SortedSet<Long> d = new TreeSet<Long>();
+			for (long value : domain) {
 				d.add(value + a);
 			}
 			// return new IntegerDomain(d);
@@ -241,13 +241,13 @@ public class IntegerDomain {
 			return d.add(lb);
 		}
 		if (domain == null || d.domain == null) {
-			int lb0 = lb + d.lb;
-			int ub0 = ub + d.ub;
+			long lb0 = lb + d.lb;
+			long ub0 = ub + d.ub;
 			return new IntegerDomain(lb0, ub0);
 		} else {
-			SortedSet<Integer> d0 = new TreeSet<Integer>();
-			for (int value1 : domain) {
-				for (int value2 : d.domain) {
+			SortedSet<Long> d0 = new TreeSet<Long>();
+			for (long value1 : domain) {
+				for (long value2 : d.domain) {
 					d0.add(value1 + value2);
 				}
 			}
@@ -256,7 +256,7 @@ public class IntegerDomain {
 		}
 	}
 
-	public IntegerDomain sub(int a) throws SugarException {
+	public IntegerDomain sub(long a) throws SugarException {
 		return add(-a);
 	}
 
@@ -264,12 +264,12 @@ public class IntegerDomain {
 		return add(d.neg());
 	}
 
-	public IntegerDomain mul(int a) throws SugarException {
+	public IntegerDomain mul(long a) throws SugarException {
 		if (domain == null) {
 			// XXX
 			if (false && size() <= MAX_SET_SIZE) {
-				SortedSet<Integer> d = new TreeSet<Integer>();
-				for (int value = lb; value <= ub; value++) {
+				SortedSet<Long> d = new TreeSet<Long>();
+				for (long value = lb; value <= ub; value++) {
 					d.add(value * a);
 				}
 				return create(d);
@@ -279,8 +279,8 @@ public class IntegerDomain {
 				return new IntegerDomain(lb*a, ub*a);
 			}
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>();
-			for (int value : domain) {
+			SortedSet<Long> d = new TreeSet<Long>();
+			for (long value : domain) {
 				d.add(value * a);
 			}
 			return create(d);
@@ -295,17 +295,17 @@ public class IntegerDomain {
 		}
 		if (domain == null || d.domain == null
 				|| size() * d.size() > MAX_SET_SIZE) {
-			int b00 = lb * d.lb;
-			int b01 = lb * d.ub;
-			int b10 = ub * d.lb;
-			int b11 = ub * d.ub;
-			int lb0 = Math.min(Math.min(b00, b01), Math.min(b10, b11));
-			int ub0 = Math.max(Math.max(b00, b01), Math.max(b10, b11));
+			long b00 = lb * d.lb;
+			long b01 = lb * d.ub;
+			long b10 = ub * d.lb;
+			long b11 = ub * d.ub;
+			long lb0 = Math.min(Math.min(b00, b01), Math.min(b10, b11));
+			long ub0 = Math.max(Math.max(b00, b01), Math.max(b10, b11));
 			return new IntegerDomain(lb0, ub0);
 		} else {
-			SortedSet<Integer> d0 = new TreeSet<Integer>();
-			for (int value1 : domain) {
-				for (int value2 : d.domain) {
+			SortedSet<Long> d0 = new TreeSet<Long>();
+			for (long value1 : domain) {
+				for (long value2 : d.domain) {
 					d0.add(value1 * value2);
 				}
 			}
@@ -313,14 +313,14 @@ public class IntegerDomain {
 		}
 	}
 
-	private int div(int x, int y) {
+	private long div(long x, long y) {
 		if (x < 0 && x % y != 0) {
 			return x / y - 1;
 		}
 		return x / y;
 	}
 
-	public IntegerDomain div(int a) throws SugarException {
+	public IntegerDomain div(long a) throws SugarException {
 		if (domain == null) {
 			if (a < 0) {
 				return new IntegerDomain(div(ub,a), div(lb,a));
@@ -328,8 +328,8 @@ public class IntegerDomain {
 				return new IntegerDomain(div(lb,a), div(ub,a));
 			}
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>();
-			for (int value : domain) {
+			SortedSet<Long> d = new TreeSet<Long>();
+			for (long value : domain) {
 				d.add(div(value, a));
 			}
 			return create(d);
@@ -342,12 +342,12 @@ public class IntegerDomain {
 		}
 		if (domain == null || d.domain == null
 				|| size() * d.size() > MAX_SET_SIZE) {
-			int b00 = div(lb, d.lb);
-			int b01 = div(lb, d.ub);
-			int b10 = div(ub, d.lb);
-			int b11 = div(ub, d.ub);
-			int lb0 = Math.min(Math.min(b00, b01), Math.min(b10, b11));
-			int ub0 = Math.max(Math.max(b00, b01), Math.max(b10, b11));
+			long b00 = div(lb, d.lb);
+			long b01 = div(lb, d.ub);
+			long b10 = div(ub, d.lb);
+			long b11 = div(ub, d.ub);
+			long lb0 = Math.min(Math.min(b00, b01), Math.min(b10, b11));
+			long ub0 = Math.max(Math.max(b00, b01), Math.max(b10, b11));
 			if (d.lb <= 1 && 1 <= d.ub) {
 				lb0 = Math.min(lb0, Math.min(lb, ub));
 				ub0 = Math.max(ub0, Math.max(lb, ub));
@@ -358,9 +358,9 @@ public class IntegerDomain {
 			}
 			return new IntegerDomain(lb0, ub0);
 		} else {
-			SortedSet<Integer> d0 = new TreeSet<Integer>();
-			for (int value1 : domain) {
-				for (int value2 : d.domain) {
+			SortedSet<Long> d0 = new TreeSet<Long>();
+			for (long value1 : domain) {
+				for (long value2 : d.domain) {
 					d0.add(div(value1, value2));
 				}
 			}
@@ -368,13 +368,13 @@ public class IntegerDomain {
 		}
 	}
 
-	public IntegerDomain mod(int a) throws SugarException {
+	public IntegerDomain mod(long a) throws SugarException {
 		a = Math.abs(a);
 		if (domain == null) {
 			return new IntegerDomain(0, a - 1);
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>();
-			for (int value : domain) {
+			SortedSet<Long> d = new TreeSet<Long>();
+			for (long value : domain) {
 				d.add(value % a);
 			}
 			return create(d);
@@ -386,13 +386,13 @@ public class IntegerDomain {
 			return mod(d.lb);
 		}
 		if (domain == null || d.domain == null) {
-			int lb0 = 0;
-			int ub0 = Math.max(Math.abs(d.lb), Math.abs(d.ub)) - 1;
+			long lb0 = 0;
+			long ub0 = Math.max(Math.abs(d.lb), Math.abs(d.ub)) - 1;
 			return new IntegerDomain(lb0, ub0);
 		} else {
-			SortedSet<Integer> d0 = new TreeSet<Integer>();
-			for (int value1 : domain) {
-				for (int value2 : d.domain) {
+			SortedSet<Long> d0 = new TreeSet<Long>();
+			for (long value1 : domain) {
+				for (long value2 : d.domain) {
 					d0.add(value1 % value2);
 				}
 			}
@@ -400,34 +400,34 @@ public class IntegerDomain {
 		}
 	}
 
-	public IntegerDomain pow(int a) throws SugarException {
+	public IntegerDomain pow(long a) throws SugarException {
 		if (domain == null) {
-			int a1 = (int)Math.round(Math.pow(lb, a));
-			int a2 = (int)Math.round(Math.pow(ub, a));
-			int lb0 = Math.min(a1, a2);
-			int ub0 = Math.max(a1, a2);
+			long a1 = (long)Math.round(Math.pow(lb, a));
+			long a2 = (long)Math.round(Math.pow(ub, a));
+			long lb0 = Math.min(a1, a2);
+			long ub0 = Math.max(a1, a2);
 			if (a % 2 == 0 && lb <= 0 && 0 <= ub) {
 				return new IntegerDomain(0, ub0);
 			} else {
 				return new IntegerDomain(lb0, ub0);
 			}
 		} else {
-			SortedSet<Integer> d = new TreeSet<Integer>();
-			for (int value : domain) {
-				d.add((int)Math.round(Math.pow(value, a)));
+			SortedSet<Long> d = new TreeSet<Long>();
+			for (long value : domain) {
+				d.add((long)Math.round(Math.pow(value, a)));
 			}
 			return create(d);
 		}
 	}
 
 	public IntegerDomain min(IntegerDomain d) throws SugarException {
-		int lb0 = Math.min(lb, d.lb);
-		int ub0 = Math.min(ub, d.ub);
+		long lb0 = Math.min(lb, d.lb);
+		long ub0 = Math.min(ub, d.ub);
 		if (ub <= d.lb) {
 			return this;
 		} else if (d.ub <= lb) {
 			return d;
-		} 
+		}
 		if (domain == null) {
 			if (d.domain == null) {
 				return new IntegerDomain(lb0, ub0);
@@ -438,7 +438,7 @@ public class IntegerDomain {
 			if (d.domain == null) {
 				return create(domain.subSet(lb0, ub0 + 1));
 			} else {
-				SortedSet<Integer> d1 = new TreeSet<Integer>(domain);
+				SortedSet<Long> d1 = new TreeSet<Long>(domain);
 				d1.addAll(d.domain);
 				d1 = d1.subSet(lb0, ub0 + 1);
 				return create(d1);
@@ -447,13 +447,13 @@ public class IntegerDomain {
 	}
 
 	public IntegerDomain max(IntegerDomain d) throws SugarException {
-		int lb0 = Math.max(lb, d.lb);
-		int ub0 = Math.max(ub, d.ub);
+		long lb0 = Math.max(lb, d.lb);
+		long ub0 = Math.max(ub, d.ub);
 		if (lb >= d.ub) {
 			return this;
 		} else if (d.lb >= ub) {
 			return d;
-		} 
+		}
 		if (domain == null) {
 			if (d.domain == null) {
 				return new IntegerDomain(lb0, ub0);
@@ -464,7 +464,7 @@ public class IntegerDomain {
 			if (d.domain == null) {
 				return create(domain.subSet(lb0, ub0 + 1));
 			} else {
-				SortedSet<Integer> d1 = new TreeSet<Integer>(domain);
+				SortedSet<Long> d1 = new TreeSet<Long>(domain);
 				d1.addAll(d.domain);
 				d1 = d1.subSet(lb0, ub0 + 1);
 				return create(d1);
@@ -472,7 +472,7 @@ public class IntegerDomain {
 		}
 	}
 
-	public SortedSet<Integer> headSet(int value) {
+	public SortedSet<Long> headSet(long value) {
 		return domain.headSet(value);
 	}
 
@@ -482,10 +482,10 @@ public class IntegerDomain {
 			                         Expression.create(ub)));
 		}
 		List<Expression> doms = new ArrayList<Expression>();
-		int value0 = Integer.MIN_VALUE;
-		int value1 = Integer.MIN_VALUE;
-		for (int value : domain) {
-			if (value0 == Integer.MIN_VALUE) {
+		long value0 = Long.MIN_VALUE;
+		long value1 = Long.MIN_VALUE;
+		for (long value : domain) {
+			if (value0 == Long.MIN_VALUE) {
 				value0 = value1 = value;
 			} else if (value1 + 1 == value) {
 				value1 = value;
@@ -513,10 +513,10 @@ public class IntegerDomain {
 			sb.append(lb + ".." + ub);
 		} else {
 			String delim = "";
-			int value0 = Integer.MIN_VALUE;
-			int value1 = Integer.MIN_VALUE;
-			for (int value : domain) {
-				if (value0 == Integer.MIN_VALUE) {
+			long value0 = Long.MIN_VALUE;
+			long value1 = Long.MIN_VALUE;
+			for (long value : domain) {
+				if (value0 == Long.MIN_VALUE) {
 					value0 = value1 = value;
 				} else if (value1 + 1 == value) {
 					value1 = value;
@@ -531,7 +531,7 @@ public class IntegerDomain {
 					value0 = value1 = value;
 				}
 			}
-			if (value0 != Integer.MIN_VALUE) {
+			if (value0 != Long.MIN_VALUE) {
 				sb.append(delim);
 				if (value0 == value1) {
 					sb.append(value0);
@@ -550,10 +550,10 @@ public class IntegerDomain {
 		} else {
 			sb.append("(");
 			String delim = "";
-			int value0 = Integer.MIN_VALUE;
-			int value1 = Integer.MIN_VALUE;
-			for (int value : domain) {
-				if (value0 == Integer.MIN_VALUE) {
+			long value0 = Long.MIN_VALUE;
+			long value1 = Long.MIN_VALUE;
+			for (long value : domain) {
+				if (value0 == Long.MIN_VALUE) {
 					value0 = value1 = value;
 				} else if (value1 + 1 == value) {
 					value1 = value;
@@ -568,7 +568,7 @@ public class IntegerDomain {
 					value0 = value1 = value;
 				}
 			}
-			if (value0 != Integer.MIN_VALUE) {
+			if (value0 != Long.MIN_VALUE) {
 				sb.append(delim);
 				if (value0 == value1) {
 					sb.append(value0);
